refactor: location service algo

+++ src/main/java/fr/formation/ecf/backend/ecf3cdabackend/vehicules/Vehicule.java
@@ -17,7 +17,7 @@ public class Vehicule extends EntityInfo {
     private String modele;
     private String immatriculation;
     private String etat;
-    private Integer prixJournee;
+    private Double prixJournee;
     private Boolean disponible;
     private String type;
